Fixed storage calculation, Adjusted entities, Removed unnecessary attributes, Added js to enter to next focusfield, Added functions to display net weight and price, Re-worked front-end to suit commissioner request, Made Empty vendor selection in selectonemenu available by adding a string in case of null object.

+++ src/main/java/hu/torma/deliveryapplication/DTO/ProductDTO.java
@@ -15,9 +15,9 @@ public class ProductDTO implements Serializable {
 
     private UnitDTO secondUnit;
 
-    private Double price;
+    private Integer price;
 
-    private Double compPercent;
+    private Integer compPercent;
 
     private String tariffnum;
