+++ src/main/java/hu/torma/deliveryapplication/service/StorageService.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 
 public interface StorageService {
     Double getSupplyOf(ProductDTO dto, SaleDTO saleDTO);
-    Boolean subQuantityOf(ProductDTO dto, Double qnt);
+    Boolean subQuantityOf(ProductDTO dto, Integer qnt);
 
     Boolean addToQuantity(ProductDTO dto);
